COM-1008 - rename downloadfile to file
