+++ src/modules/client/components/contracts/Contracts.vue
@@ -92,7 +92,7 @@ import { Cookies } from 'quasar';
 import esign from '@api/Esign.js';
 import { NotifyNegative } from '@components/popup/notify.js';
 import ResponsiveTable from '@components/table/ResponsiveTable';
-import { downloadDocxFile } from '@helpers/downloadFile';
+import { downloadDocxFile } from '@helpers/file';
 import { formatIdentity } from '@helpers/utils';
 import { CONTRACT_STATUS_OPTIONS, CUSTOMER_CONTRACT, COACH, CUSTOMER, AUXILIARY, COMPANY_CONTRACT } from 'src/core/data/constants.js';
 import { generateContractFields } from 'src/modules/client/helpers/generateContractFields';
